Implement new login page and authentication guards
This commit implements a new login page with its respective HTML, CSS, and TypeScript files. It utilizes the ReactiveFormsModule for handling form inputs and the AuthService for the login functionality. Furthermore, this update includes the setting up of the authentication guards that redirect users based on their logged-in status. Changes were also made to the app-wide routes to accommodate these new additions. As a precautionary measure, the localStorage service was updated to account for platform differences.